Create site.js
This is a graphical representation of the skills that I have accumulated. The inside ring separates back end from front end skills and subsequent rings two and three illustrate where these skills were acquired via my industry work placements, during my time as a lecturer and while undertaking my MSc. The skills are as spokes from the wheel with the height indicating ability, skills are according to the inner ring and colour back or front end categories.
    
In the design of the skills wheel per quater was intended to contain an equal number of 4 slices 
(4 * 1 Slice at 22.5 degrees = 90 deg angle) with a total of 16 slices for the entire wheel
therefore for simplicity sake there can only be a nominated 16 skills in total,
the animated wheel is not adaptable to more slices without manual recalculation. 

Each slice of the pie projects out - to be associated with inner rings at the point of crossing and the level of skill proficiency indicated by the height of the slice

The 0/360 degrees point of the circle starts at the 3 o'clock position and moves in a positive direction to the 6 o'clock position which is 90 degrees, the 9 o'clock position which is 180 degrees, the 12 o'clock position which is 270 degrees and finally back to the 3 o'clock position at 360 degrees.